♻️ refactor: socket관련 파일명 변경
